evdevs.c: Fix failure to read evtype on big endian
The EVIOCGBIT ioctl actually doesn't give you the bits as a sequence
of uint8_t:s, but rather as a sequence of long:s.  This makes a
difference on big endian since the bits within these longs are stored
in little endian order (i.e. event type 0 is represented by the _least_
significant bit of the first long), so we need to know where the
discontinuities are...